+++ packages/flower-core/src/interfaces/index.ts
@@ -1,6 +1,5 @@
 export * from './CoreInterface'
 export * from './ReducerInterface'
 export * from './SelectorsInterface'
-// @ts-expect-error error name types
 export * from './Store'
 export * from './UtilsInterface'