+++ packages/flower-demo/src/Examples/Example9.tsx
@@ -5,7 +5,8 @@ import {
   FlowerAction,
   useFlower,
   useFlowerForm,
-  Flower
+  Flower,
+  FlowerRule,
 } from '@flowerforce/flower-react'
 import { useEffect } from 'react'
 import './styles.css'
@@ -37,13 +38,35 @@ export function Example9() {
               id="username"
               validate={[
                 {
-                  rules: { $and: [{ username: { $exists: true } }] },
-                  message: 'Field is required'
+                  rules: {
+                    $and: [
+                      {
+                        username: {
+                          $exists: true,
+                        },
+                      },
+                    ],
+                  },
+                  message: 'Field is required',
                 },
                 {
-                  rules: { $and: [{ username: { $strGte: '6' } }] },
-                  message: 'Field length must be greater than or equal to 6.'
-                }
+                  rules: {
+                    $and: [
+                      {
+                        username: {
+                          $strGte: '6',
+                        },
+                      },
+                    ],
+                  },
+                  message: 'Field length must be greater than or equal to 6.',
+                },
+                {
+                  rules: (data: any) => {
+                    return true
+                  },
+                  message: 'Error custom',
+                },
               ]}
             >
               {({ onChange, value = '', errors, onBlur, hidden }) => (
@@ -62,6 +85,13 @@ export function Example9() {
                 </div>
               )}
             </FlowerField>
+            <FlowerRule
+              rules={(data: any) => {
+                return !data.$form.isValid
+              }}
+            >
+              Ciao Andrea
+            </FlowerRule>
           </div>
           <div className="field">
             <label htmlFor="password">Password *</label>
@@ -70,8 +100,8 @@ export function Example9() {
               validate={[
                 {
                   rules: { $and: [{ password: { $exists: true } }] },
-                  message: 'Field is required'
-                }
+                  message: 'Field is required',
+                },
               ]}
             >
               {({ onChange, value = '', errors, onBlur, hidden }) => (