+++ packages/flower-react/src/components/useFlower.ts
@@ -3,6 +3,8 @@ import { context } from '../context'
 import { makeSelectCurrentNodeId } from '../selectors'
 import { useDispatch, useSelector } from '../provider'
 import { UseFlower } from './types/FlowerHooks'
+import { Emitter } from '@flowerforce/flower-core'
+import _get from 'lodash/get'
 
 type NavigateFunctionParams = string | Record<string, any>
 
@@ -95,16 +97,16 @@ const useFlower: UseFlower = ({ flowName: customFlowName, name } = {}) => {
     (params: any) => {
       /* istanbul ignore next */
       // eslint-disable-next-line no-underscore-dangle, no-undef
-      // if (global.window && global.window.__FLOWER_DEVTOOLS__) {
-      //   Emitter.emit('flower-devtool-from-client', {
-      //     source: 'flower-client',
-      //     action: 'FLOWER_NAVIGATE',
-      //     nodeId,
-      //     name: flowName,
-      //     time: new Date(),
-      //     params,
-      //   });
-      // }
+      if (_get(global.window, '__FLOWER_DEVTOOLS__')) {
+        Emitter.emit('flower-devtool-from-client', {
+          source: 'flower-client',
+          action: 'FLOWER_NAVIGATE',
+          nodeId,
+          name: flowName,
+          time: new Date(),
+          params
+        })
+      }
     },
     [flowName, nodeId]
   )