+++ packages/flower-core/CHANGELOG.md
@@ -1,3 +1,21 @@
+## 3.1.0 (2024-06-06)
+
+
+### 🚀 Features
+
+- add restart action in core ([800da33](https://github.com/flowerforce/flower/commit/800da33))
+
+- add devtool to demo - remove duplicated export type ([88117d6](https://github.com/flowerforce/flower/commit/88117d6))
+
+- allow custom functions when defining rules and navigation between nodes ([b35f5da](https://github.com/flowerforce/flower/commit/b35f5da))
+
+
+### 🩹 Fixes
+
+- fix circular dependency on rules matchers ([0bcb02f](https://github.com/flowerforce/flower/commit/0bcb02f))
+
+- align packages ([210a284](https://github.com/flowerforce/flower/commit/210a284))
+
 ## 3.0.0 (2024-05-20)
 
 This was a version bump only for flower-core to align it with other projects, there were no code changes.